requirements.txt
streamlit
azure-storage-blob
trimesh
matplotlib
numpy-stl
cryptography
ipyvolume
pyarrow
ipython_genutils
fastapi
uvicorn
python-dotenv
python-multipart
scipy
numpy
scikit-learn
requests
werkzeug
psycopg2-binary
watchdog